+++ flash-graph/matrix/min-tri-kmeans.cpp
@@ -290,7 +290,7 @@ namespace fg
 {
     unsigned compute_min_kmeans(const double* matrix, double* clusters_ptr,
             unsigned* cluster_assignments, unsigned* cluster_assignment_counts,
-            const unsigned num_rows, const unsigned num_cols, const size_t k,
+            const size_t num_rows, const size_t num_cols, const unsigned k,
             const unsigned MAX_ITERS, const int max_threads, const std::string init,
             const double tolerance, const std::string dist_type)
     {
